reorganise selection code
