+++ src/table/selection/__tests__/use-selection.test.tsx
@@ -1,8 +1,8 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { useSelection } from '../../../lib/components/table/use-selection';
-import { renderHook } from '../../__tests__/render-hook';
+import { useSelection } from '../../../../lib/components/table/selection';
+import { renderHook } from '../../../__tests__/render-hook';
 
 describe('useSelection', () => {
   it('satisfies istanbul coverage', () => {
