+++ src/table/selection/index.ts
@@ -1,0 +1,8 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export { SelectionProps } from './interfaces';
+export { SelectionControl, SelectionControlProps } from './selection-control';
+export { useSelectionFocusMove } from './use-selection-focus-move';
+export { useSelection } from './use-selection';
+export { focusMarkers } from './utils';
