+++ src/table/selection/styles.scss
